Move stream-related extension methods to StreamGroovyMethods
