+++ build.gradle
@@ -231,10 +231,10 @@ tasks.named('test') {
 
 if (!JavaVersion.current().java10Compatible) {
     logger.lifecycle '''
-    **************************************** WARNING ********************************************
-    ******   You are running the build with an older JDK. NEVER try to release with 1.8.   ******
-    ******   You must use a JDK 10+ in order to compile all features of the language.     ******
-    *********************************************************************************************
+**************************************** WARNING ********************************************
+******   You are running the build with an older JDK. NEVER try to release with 1.8.   ******
+******   You must use a JDK 10+ in order to compile all features of the language.      ******
+*********************************************************************************************
 '''
 }
